Add binoculars mod with binoculars item to allow zoom
Uses the new player object property 'can zoom'.
In survival mode, use of zoom requires the binoculars item in
inventory.
Zoom is automatically allowed in creative mode and for players with
the 'creative' privilege.
The binoculars.update_player_property() function is global so can be
redefined by a mod for alternative behaviour.